--- moz_do/__init__.py
@@ -1,4 +1,0 @@
-from flask import Flask
-moz_do = Flask(__name__)
-
-from moz_do import views
