+++ moz_do/website/__init__.py
@@ -1,0 +1,4 @@
+from flask import Flask
+moz_do = Flask(__name__)
+
+from moz_do.website import views
